Changing routes to have multiple services and controllers within the same app.
