+++ personna/controllers/admin/equipmentController.js
