+++ personna/controllers/admin/exerciseEquipmentController.js
