+++ personna/services/admin/exerciseEquipmentService.js
@@ -5,7 +5,7 @@
 "use strict"
 // Imports
 const Q                 = require('q'),
-      BaseService       = require('./baseService').BaseService,
+      BaseService       = require('../baseService').BaseService,
       util              = require('util');
 // filtered query filters
 const eEquipPaths  = {
